[FEAT] 시험관련 API Service구현(#12)
