+++ LevelUpiOS/LevelUpiOS/Network/Base/HTTPHeaderField.swift
@@ -14,6 +14,6 @@ enum HTTPHeaderField: String {
 }
 
 enum ContentType: String {
-    case json = "Application/json"
+    case json = "application/json"
 }
