+++ LevelUpiOS/LevelUpiOS/Network/Base/TargetType.swift
@@ -31,8 +31,6 @@ extension TargetType {
             method: method
         )
 
-        urlRequest.setValue(ContentType.json.rawValue, forHTTPHeaderField: HTTPHeaderField.contentType.rawValue)
-
         switch parameters {
         case .requestWithBody(let request):
             let params = request?.toDictionary() ?? [:]
